chore: move isEstimatedReturnLow logic to util file
