+++ ui/pages/swaps/prepare-swap-page/review-quote.js
@@ -44,6 +44,7 @@ import {
   fetchSwapsSmartTransactionFees,
   getSmartTransactionFees,
   getCurrentSmartTransactionsEnabled,
+  getIsEstimatedReturnLow,
 } from '../../../ducks/swaps/swaps';
 import { getCurrentChainId } from '../../../../shared/modules/selectors/networks';
 import {
@@ -97,7 +98,6 @@ import {
   SLIPPAGE_HIGH_ERROR,
   SLIPPAGE_LOW_ERROR,
   MAX_ALLOWED_SLIPPAGE,
-  SWAPS_QUOTE_MAX_RETURN_DIFFERENCE_PERCENTAGE,
 } from '../../../../shared/constants/swaps';
 import { GasRecommendations } from '../../../../shared/constants/gas';
 import CountdownTimer from '../countdown-timer';
@@ -149,7 +149,6 @@ import { getTokenFiatAmount } from '../../../helpers/utils/token-util';
 import { toChecksumHexAddress } from '../../../../shared/modules/hexstring-utils';
 import { useAsyncResult } from '../../../hooks/useAsyncResult';
 import { useGasFeeEstimates } from '../../../hooks/useGasFeeEstimates';
-import { useTokenFiatAmount } from '../../../hooks/useTokenFiatAmount';
 import ViewQuotePriceDifference from './view-quote-price-difference';
 import SlippageNotificationModal from './slippage-notification-modal';
 
@@ -1128,50 +1127,10 @@ export default function ReviewQuote({
     currentCurrency,
   ]);
 
-  const sourceTokenAmount = calcTokenAmount(
-    usedQuote?.sourceAmount,
-    usedQuote?.sourceTokenInfo?.decimals,
-  );
-  const sourceTokenFiatAmount = useTokenFiatAmount(
-    usedQuote?.sourceTokenInfo?.address,
-    sourceTokenAmount || 0,
-    usedQuote?.sourceTokenInfo?.symbol,
-    {
-      showFiat: true,
-    },
-    true,
-    null,
-    false,
-  );
-  const destinationTokenAmount = calcTokenAmount(
-    usedQuote?.destinationAmount,
-    usedQuote?.destinationTokenInfo?.decimals,
-  );
-  const destinationTokenFiatAmount = useTokenFiatAmount(
-    usedQuote?.destinationTokenInfo?.address,
-    destinationTokenAmount || 0,
-    usedQuote?.destinationTokenInfo?.symbol,
-    {
-      showFiat: true,
-    },
-    true,
-    null,
-    false,
-  );
-  const adjustedReturnValue =
-    destinationTokenFiatAmount && rawNetworkFees
-      ? new BigNumber(destinationTokenFiatAmount).minus(
-          new BigNumber(rawNetworkFees),
-        )
-      : null;
-  const isEstimatedReturnLow =
-    sourceTokenFiatAmount && adjustedReturnValue
-      ? adjustedReturnValue.lt(
-          new BigNumber(sourceTokenFiatAmount).times(
-            1 - SWAPS_QUOTE_MAX_RETURN_DIFFERENCE_PERCENTAGE,
-          ),
-        )
-      : false;
+  const isEstimatedReturnLow = getIsEstimatedReturnLow({
+    usedQuote,
+    rawNetworkFees,
+  });
   setIsEstimatedReturnLow(isEstimatedReturnLow);
 
   return (
